requirements.txt
requests
streamlit
google-genai
